feat: Manage Events - track app lifecycle events (#118)
* move all the logic inside confidence

* add evaluation extension in flag resolution

* cleanup

* have the evaluation in the confidence module for the flags

* update demo app to support only confidence

* fix some tests to move to confidence

* use int instead of int64 for 32bits system to work with default value for int, fix some more tests

* fixup! use int instead of int64 for 32bits system to work with default value for int, fix some more tests

* fixup! Merge branch 'main' into move-flag-evaluation-confidence

* fixup! fixup! Merge branch 'main' into move-flag-evaluation-confidence

* fixup! fixup! fixup! Merge branch 'main' into move-flag-evaluation-confidence

* add analytics for app and ui kit lifecycle, add the appear and disappear for demo app events

* fixup! add analytics for app and ui kit lifecycle, add the appear and disappear for demo app events

* handle app launch and app install and app updates

* add context producer and produce context for is_foreground and build and version

* fixup! merge main

* fixup! fixup! merge main

* refactor: Smaller refactor and fixes

* Remove some managed events

* wip: minor fixes and experiments

* introduce passthrough subject with buffer

* update the engine to remove the batches by default and add them if failed

* fixup! update the engine to remove the batches by default and add them if failed

* queue label prefix

* use semaphore to make upload serialz

* Remove dead code

* Make buffered passthrough serial

---------

+++ Sources/Confidence/EventStorage.swift
@@ -95,7 +95,9 @@ internal class EventStorageImpl: EventStorage {
     func remove(id: String) throws {
         try storageQueue.sync {
             let fileUrl = folderURL.appendingPathComponent(id)
-            try FileManager.default.removeItem(at: fileUrl)
+            if FileManager.default.fileExists(atPath: fileUrl.path) {
+                try FileManager.default.removeItem(at: fileUrl)
+            }
         }
     }
 
@@ -114,7 +116,7 @@ internal class EventStorageImpl: EventStorage {
             self.currentFileHandle = try FileHandle(forWritingTo: currentFile)
         } else {
             // Create a brand new file
-            let fileUrl = folderURL.appendingPathComponent(String(Date().timeIntervalSince1970))
+            let fileUrl = folderURL.appendingPathComponent(String(UUID().uuidString))
             FileManager.default.createFile(atPath: fileUrl.path, contents: nil)
             self.currentFileUrl = fileUrl
             self.currentFileHandle = try FileHandle(forWritingTo: fileUrl)
